Fixing scripts to v3.3.3

+++ iguana/next-blazegraph.sh
@@ -1,23 +1,26 @@
 graphFolder=$1
-for f in $(ls $graphFolder)
+for f in "$graphFolder"/*
 do
-	fullFile=$graphFolder$f
-	rm -rf queryInstances/
-	echo "executing $f"
-	pkill -f blazegraph
-	./start-ref.sh $fullFile
-	./start-blazegraph.sh $fullFile
-	sleep 5m
-	#gn=$(echo ${f:43} | sed -e 's/[^a-zA-Z0-9]//g')
-	#gn="geology_target"
-	gn="${f%%.*}"
-	echo "$gn"
-	cp example-suite.yml example-suiteTMP.yml
-        sed -i -e 's,GraphName,'"$gn"',g' example-suiteTMP.yml
-        sed -i -e 's,TSTORE,Blazegraph,g' example-suiteTMP.yml
-        sed -i -e 's,ENDPOINT,http://localhost:9999/blazegraph/sparql,g' example-suiteTMP.yml
-        ./start-iguana.sh example-suiteTMP.yml
+  if [ -f "$f" ]; then
+    rm -rf queryCache/
+    echo "executing $f"
+    pkill -f blazegraph
+    ./start-ref.sh $f
+    ./start-blazegraph.sh $f
+    sleep 5m
+    gn="${f%%.*}"
+    echo "$gn"
+    cp example-suite.yml example-suiteTMP.yml
+    queries=$( ./get_queries.sh $f )
+    echo "Queries file: $queries"
+    sed -i -e 's,QUERIES,'"$queries"',g' example-suiteTMP.yml
+    sed -i -e 's,GraphName,'"$gn"',g' example-suiteTMP.yml
+    sed -i -e 's,TSTORE,Blazegraph,g' example-suiteTMP.yml
+    sed -i -e 's,ENDPOINT,http://localhost:9999/blazegraph/sparql,g' example-suiteTMP.yml
+    ./start-iguana.sh example-suiteTMP.yml
 
-	pkill -f blazegraph
-	echo "done"
+    pkill -f blazegraph
+  fi
+  echo "done"
 done
+
