+++ iguana/next-tentris.sh
@@ -1,20 +1,26 @@
 graphFolder=$1
-for f in $(ls $graphFolder)
+for f in "$graphFolder"/*
 do
-	fullFile=$graphFolder$f
-	rm -rf queryInstances/
-	./start-ref.sh $fullFile
-	echo "executing $f"
-	pkill -f tentris
-	./start-tentris.sh $graphFolder $f
-	sleep 5m
-	gn="${f%%.*}"
-	echo "$gn"
-	cp example-suite.yml example-suiteTMP.yml
-	sed -i -e 's,GraphName,'"$gn"',g' example-suiteTMP.yml
-	sed -i -e 's,TSTORE,Tentris,g' example-suiteTMP.yml
-	sed -i -e 's,ENDPOINT,http://localhost:9080/sparql,g' example-suiteTMP.yml
-	./start-iguana.sh example-suiteTMP.yml
-	pkill -f tentris
-	echo "done"
+  if [ -f "$f" ]; then
+    rm -rf queryCache/
+    echo "executing $f"
+    pkill -f tentris
+    ./start-ref.sh $f
+    ./start-tentris.sh $f
+    sleep 5m
+    gn="${f%%.*}"
+    echo "$gn"
+    cp example-suite.yml example-suiteTMP.yml
+    queries=$( ./get_queries.sh $f )
+    echo "Queries file: $queries"
+    sed -i -e 's,QUERIES,'"$queries"',g' example-suiteTMP.yml
+    sed -i -e 's,GraphName,'"$gn"',g' example-suiteTMP.yml
+    sed -i -e 's,TSTORE,Tentris,g' example-suiteTMP.yml
+    sed -i -e 's,ENDPOINT,http://localhost:9080/sparql,g' example-suiteTMP.yml
+    ./start-iguana.sh example-suiteTMP.yml
+
+    pkill -f tentris
+  fi
+  echo "done"
 done
+
